Add flag to prevent interrupt disabling, fix bad data after read failure.
Added the ability to prevent disabling of interrupts. This causes reads to fail more
frequently than with interrupts, but is required to prevent interference with other
code that needs interrupts,for example AC phase control.

Fixed a bug that caused the incorrect data to be returned for the rest of the
poll period after a read failed. If a read failed then that call would correctly
return NAN but subsequent calls would return whatever was in the buffer.